+++ app/routes/components/Select/changelog.codegen.ts
@@ -21,4 +21,5 @@ export const changelog = ''
   + '- 2025-11-21: Add `width` prop to control the width of the component. Remove the\n'
   + '  `min-width` constraint on the trigger. And remove the dependency on\n'
   + '  `ThemeProvider`.\n'
+  + '- 2025-11-24: Support the empty state.\n'
   ;